refactor: 도메인별 분리 적용
