+++ src/advanced/components/cart/index.ts
@@ -1,0 +1,3 @@
+// Cart domain components
+export { default as CartContainer } from './CartContainer';
+export { default as CartItem } from './CartItem';
