+++ src/advanced/components/common/Toast/ToastContainer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Toast from './Toast';
-import { useToast } from '../../contexts/ToastContext';
+import { useToast } from '../../../contexts/ToastContext';
 
 const ToastContainer: React.FC = () => {
   const { toasts, removeToast } = useToast();
