+++ src/advanced/components/common/index.ts
@@ -1,0 +1,5 @@
+// Common UI components
+export { default as Header } from './Header';
+export { default as DiscountIcon } from './DiscountIcon';
+export { default as ToastContainer } from './Toast/ToastContainer';
+export { default as Toast } from './Toast/Toast';
