+++ src/advanced/components/help/index.ts
@@ -1,0 +1,2 @@
+// Help domain components
+export { default as HelpModal } from './HelpModal';
