+++ src/advanced/components/order/index.ts
@@ -1,0 +1,2 @@
+// Order domain components
+export { default as OrderSummary } from './OrderSummary';
