+++ src/advanced/components/product/index.ts
@@ -1,0 +1,3 @@
+// Product domain components
+export { default as ProductSelector } from './ProductSelector';
+export { default as StockStatus } from './StockStatus';
